backend/requirements.txt
annotated-doc==0.0.4
motor>=3.3.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
google-auth==2.48.0
google-auth-oauthlib==1.3.0
google-genai==1.66.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
oauthlib==3.3.1
packaging==26.0
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
pypdf==6.7.5
pypdfium2==5.5.0
pytesseract==0.3.13
python-dotenv==1.2.2
python-multipart==0.0.22
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
